refactor: extract remaining module tests into _test.zig files
Covers ChunkedLeaf, slot_math, any_beacon_state, ThreadPool, era and
bounded_array, and renames swap_or_not_shuffle/test.zig to root_test.zig
so every test file follows the same <module>_test.zig convention.

Moves the pre-existing node/proof/view/clock test imports off the package
roots and onto the modules they test. The memory_safety_test.zig files
stay wired at their package root: they are module-wide by nature, not
paired with one module.

Also makes test discovery explicit in persistent_merkle_tree/root.zig and
ssz/root.zig rather than relying on refAllDecls reaching files through the
re-export list, which is how tree_view/bit_list.zig's tests went missing.
