+++ src/swap_or_not_shuffle/root.zig
@@ -600,5 +600,5 @@ pub fn computePtcIndicesForEpochInto(
 }
 
 test {
-    _ = @import("./test.zig");
+    _ = @import("root_test.zig");
 }
